fix: Address clang-tidy warnings
- consensus.c: Fix narrowing conversion in popcount32()
- onion.c: Add NOLINT for intentional non-null-terminated memcpy (binary data)
- storage.c: Add CYXWIZ_POS_FAIL_NONE enum value for success case
- lora.c: Fix switch fallthrough, narrowing conversion, add NOLINT for
  non-cryptographic rand() usage and false positive analyzer warnings

+++ src/core/consensus.c
@@ -162,7 +162,7 @@ static int popcount32(uint32_t x)
 {
     int count = 0;
     while (x) {
-        count += x & 1;
+        count += (int)(x & 1);
         x >>= 1;
     }
     return count;
